requirements.txt
regex
PyYAML
pandas
SQLAlchemy
sklearn
scikit-learn
Scrapy
itemloaders
w3lib
itemadapter
vncorenlp
transformers
